deleted database.h
