+++ src/main.cpp
@@ -23,7 +23,6 @@
 #include <filesystem>
 #include "FileRead.h"
 #include "RegEx.h"
-#include "Database.h"
 
 using namespace std;
